TimeInput TS
